backend/requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
asgiref==3.7.2
attrs==23.2.0
azure-ai-formrecognizer==3.3.2
azure-common==1.1.28
azure-core==1.30.1
beautifulsoup4==4.12.3
blinker==1.7.0
cachetools==5.3.2
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
commandlines==0.4.1
contourpy==1.2.0
cycler==0.12.1
decorator==4.4.2
filelock==3.13.1
filetype==1.2.0
Flask==3.0.2
Flask-Cors==4.0.0
fonttools==4.49.0
frozenlist==1.4.1
fsspec==2024.2.0
future==1.0.0
google==3.0.0
google-ai-generativelanguage==0.4.0
google-api-core==2.17.1
google-auth==2.28.1
google-generativeai==0.3.2
googleapis-common-protos==1.62.0
grpcio==1.62.0
grpcio-status==1.62.0
hsh==1.1.0
idna==3.6
imageio-ffmpeg==0.4.9
isodate==0.6.1
itsdangerous==2.1.2
Jinja2==3.1.3
kiwisolver==1.4.5
lazy_loader==0.3
llvmlite==0.42.0
MarkupSafe==2.1.5
matplotlib==3.8.3
more-itertools==10.2.0
mpmath==1.3.0
msrest==0.7.1
multidict==6.0.5
networkx==3.2.1
numpy==1.26.4
oauthlib==3.2.2
packaging==23.2
pandas==2.2.1
pillow==10.2.0
proto-plus==1.23.0
protobuf==4.25.3
pyasn1==0.5.1
pyasn1-modules==0.3.0
PyMuPDF==1.23.25
PyMuPDFb==1.23.22
pyparsing==3.1.1
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2024.1
requests==2.31.0
requests-oauthlib==1.4.0
rsa==4.9
six==1.16.0
soupsieve==2.5
sympy==1.12
tqdm==4.66.2
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
waitress==3.0.0
Werkzeug==3.0.1
yarl==1.9.4
